requirements.txt
fastapi
uvicorn

pymongo
motor

python-dotenv